🚚 Rename: 북마크한 공고 관련 폴더 생성 (#231)
