+++ src/components/my-page/index.tsx
@@ -1,6 +1,6 @@
 import MyPageSideBar from '@/components/my-page/side-bar/MyPageSideBar'
 import { UserInfoDescription } from '@/components/my-page/my-info/InfoDescription'
-import BookmarkedRecruitmentCard from '@/components/my-page/BookmarkedRecruitmentCard'
+import BookmarkedRecruitmentCard from '@/components/my-page/bookmarked-recruitment/BookmarkedRecruitmentCard'
 import { PasswordChange } from '@/components/my-page/my-info/PasswordChange'
 import { Withdrawal } from '@/components/my-page/my-info/Withdrawal'
 import { InfoUpdate } from '@/components/my-page/my-info/InfoUpdate'
